+++ src/kit/Changelog.md
@@ -1,10 +1,16 @@
 # Changelog
 
+## Version 0.5.4 (work in progress)
+* Add RadioButtons class
+* Improve Checkbox class
+* Global improvements
+
 ## Version 0.5.3
 * Add TextFields class (filledDraw only)
 * Improve theme (in all ui components) to match with Material Design
 * Add dynamic font imports for texts (fix size issue & load only fonts needed)
 * Add missing function in UiTexts class
+* Improve IME functions
 * Add beforeEnter and beforeExit functions in View triggered in view change
 * Add text fields sample